Use BigInt for ASN1 schema constants
Removes need for `integerValue` and `unsignedIntegerValue` extensions on BigInt which created issues with building for all platforms.

+++ Sources/PotentASN1/Schema.swift
@@ -60,7 +60,7 @@ public indirect enum Schema {
   case dynamic(unknownTypeSchema: Schema? = nil, DynamicMap)
 
   case version(Schema)
-  case versioned(range: ClosedRange<UInt>, Schema)
+  case versioned(range: ClosedRange<BigInt>, Schema)
 
   case optional(Schema)
 
@@ -68,7 +68,7 @@ public indirect enum Schema {
   case explicit(UInt8, in: ASN1.Tag.Class = .contextSpecific, Schema)
 
   case boolean(default: Bool? = nil)
-  case integer(allowed: Swift.Range<Int>? = nil, default: Int? = nil)
+  case integer(allowed: Swift.Range<BigInt>? = nil, default: BigInt? = nil)
   case real
   case bitString(size: Size? = nil)
   case octetString(size: Size? = nil)
